dev-requirements.txt
attrs==20.1.0
bokeh==2.1.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
click-log==0.3.2
cloudpickle==1.5.0
contextvars==2.4
coverage==5.2.1
cycler==0.10.0
dask==2.23.0
databricks-cli==0.11.0
decorator==4.4.2
distributed==2.23.0
dotty-dict==1.2.1
fsspec==0.8.0
gitdb==4.0.5
GitPython==3.1.7
HeapDict==1.0.1
idna==2.9
immutables==0.14
importlib-metadata==1.7.0
iniconfig==1.0.1
Jinja2==2.11.2
kiwisolver==1.2.0
locket==0.2.0
MarkupSafe==1.1.1
matplotlib==3.3.1
more-itertools==8.4.0
msgpack==1.0.0
munch==2.5.0
networkx==2.4
numpy==1.19.1
packaging==20.4
pandas==1.1.1
partd==1.1.0
Pillow==7.2.0
pluggy==0.13.1
psutil==5.7.2
py==1.9.0
pygrok==1.0.0
pyparsing==2.4.7
pytest==6.0.1
pytest-asyncio==0.14.0
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
regex==2020.7.14
requests==2.24.0
setuptools-scm==4.1.2
six==1.15.0
smmap==3.0.4
sortedcontainers==2.2.2
streamz==0.5.5
tabulate==0.8.7
tblib==1.7.0
tenacity==6.2.0
toml==0.10.1
toolz==0.10.0
tornado==6.0.4
typing-extensions==3.7.4.2
urllib3==1.25.9
zict==2.0.0
zipp==3.1.0
